+++ src/settingsObjects/deviceSettings/dachsSettings.ts
@@ -39,6 +39,9 @@ export class DachsDeviceSettings extends ActuatorSettings implements iDachsDevic
   public warmWaterDesiredMaxTemp: number = 75;
 
   /** @inheritDoc */
+  public heatStorageMaxStartTemp: number = 70;
+
+  /** @inheritDoc */
   public winterMinimumHeatStorageTemp: number = 55;
   /** @inheritDoc */
   public winterMinimumPreNightHeatStorageTemp: number = 65;
@@ -59,6 +62,7 @@ export class DachsDeviceSettings extends ActuatorSettings implements iDachsDevic
     this.batteryLevelAllowStartThreshold = data.batteryLevelAllowStartThreshold ?? this.batteryLevelAllowStartThreshold;
     this.warmWaterDesiredMinTemp = data.warmWaterDesiredMinTemp ?? this.warmWaterDesiredMinTemp;
     this.warmWaterDesiredMaxTemp = data.warmWaterDesiredMaxTemp ?? this.warmWaterDesiredMaxTemp;
+    this.heatStorageMaxStartTemp = data.heatStorageMaxStartTemp ?? this.heatStorageMaxStartTemp;
     this.winterMinimumHeatStorageTemp = data.winterMinimumHeatStorageTemp ?? this.winterMinimumHeatStorageTemp;
     this.winterMinimumPreNightHeatStorageTemp =
       data.winterMinimumPreNightHeatStorageTemp ?? this.winterMinimumPreNightHeatStorageTemp;